relegated some of the ticket adding logic to the repository
