+++ TickAPI/TickAPI/ShoppingCarts/Abstractions/IShoppingCartRepository.cs
@@ -8,4 +8,5 @@ public interface IShoppingCartRepository
 {
     public Task<Result<ShoppingCart>> GetShoppingCartByEmailAsync(string customerEmail);
     public Task<Result> UpdateShoppingCartAsync(string customerEmail, ShoppingCart shoppingCart);
+    public Task<Result> AddNewTicketToCartAsync(string customerEmail, Guid ticketTypeId, uint amount);
 }
